replace writeEnv with GatsbyAdapter.initEnv
